OCR_DATA_IN_DB/requirements.txt
datasets==1.18.3
easygui==0.98.2
JSON_log_formatter==0.5.1
numpy==1.22.2
opencv_python==4.5.5.62
pandas==1.4.1
pdf2image==1.16.0
Pillow==9.2.0
PyPDF2==2.10.8
pytesseract==0.3.9
requests==2.27.1
scipy==1.8.0
torch==1.10.2
transformers==4.16.2
